Moteur recommandation/requirements.txt
flask
flask-cors
pandas
fpdf2
